requirements.txt
ollama>=0.5.1
pydantic>=2.8.2
python-dotenv>=1.0.1
